% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{aci}
\alias{aci}
\title{Calculating the Advocacy Coalition Index.}
\usage{
aci(matrix, policy.score, cut.value = 0.5)
}
\arguments{
\item{matrix}{a matrix.}

\item{policy.score}{a vector indicating the policy score ranging 0 to 1.}

\item{cut.value}{the cut-off value to differentiate the homopilious ties. The
default value is 0.5.}
}
\value{
a list object containing,

(1) actor level broker-ties, missing-ties;score as well as ACI (in-coming, and out-going score differentiated);

(2) the whole network level ACI, broker ties score and missing-ties score;

(3) coalition matrix;

(4) original matrix supplied as matrix;

(5) preference agreement matrix.
}
\description{
\code{aci} calculates the advocacy coalition index.
}
\details{
This function returns the adovcacy coalition index (ACI) and homophily matrix.
}
\examples{
# matrix
mat <- matrix(c(0,1,1,0,0,0,
                1,0,0,0,0,0,
                1,0,0,0,0,1,
                0,0,0,0,1,1,
                0,0,0,1,0,1,
                0,0,1,1,1,0),6,6,byrow = T)

namelist <- c(paste(rep("A",3),1:3, sep = ""),
              paste(rep("B",3),1:3, sep = ""))
dimnames(mat)<-list(namelist,namelist)

# policy score
policy.score <- c(1,1,1,0,0,0)
names(policy.score) <- namelist

# ACI
ACI <- aci(matrix = mat,                # imput matrix
             policy.score = policy.score, # imput policy score
             cut.value = 0.5)             # cut-off value (alpha) for the coalition matrix

ACI$act                 # actor-level ACI
ACI$whole               # network-level  ACI
ACI$coalition.mat       # coalition matrix
ACI$original.matrix.Y
ACI$preference.agreement.matrix.X

mat2 <- mat
mat2[2,3]<-1;mat2[3,2]<-1
mat2[3,6]<-0;mat2[6,3]<-0

aci(mat2,policy.score)

Example 3: completely opposite to ACF
mat3 <- mat2
mat3 <- 1-mat3
aci(mat3,policy.score)

}
\references{
Satoh, Keiichi, Antti Gronow, and Tuomas Yla-Anttila. (2019). "The Advocacy Coalition Index: Identifying Coalitions by Simultaneously Taking into Account Coordination of Action and Belief Homophily". Paper Presented at ECPR 2019.(\url{https://ecpr.eu/Events/PaperDetails.aspx?PaperID=46846&EventID=123}).
}
