% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr2mat.R
\name{attr2mat}
\alias{attr2mat}
\title{Convert attribute vector to a matrix}
\usage{
attr2mat(attr, method = "match", diag.val = 0)
}
\arguments{
\item{attr}{vector of input attribute.}

\item{method}{\code{"match"}, \code{"match.one"}, \code{"diff"},
\code{"abs.diff"}, \code{"sqd.diff"}, \code{"sum"}, \code{"product"},
\code{"identity"}, \code{"row"}, \code{"column"}, \code{"max"},
\code{"min"} or \code{"mean"}. Default is \code{"match"}.}

\item{diag.val}{the diag value to be filled. If \code{NULL} is set, no value
is inputed.}
}
\value{
a matrix converted from a vector.
}
\description{
\code{attr2mat} converts a vector indicating attribute to a matrix.
}
\details{
This function convertes a vector of matrices \eqn{attr} to a new matrix
\eqn{mat[i,j]} based on the definition set by \code{method}:

\code{"match"}(matching the value): \eqn{Matrix[i,j] = 1} if \eqn{attr[i] = attr[j]}; 0 otherwise.

\code{"match.one"}(maching the value 1): \eqn{Matrix[i,j] = 1} if \eqn{attr[i] = attr[j] = 1}; 0 otherwise.

\code{"diff"}(difference): \eqn{Matrix[i,j] = attr[i] - [j]}.

\code{"abs.diff"}(absolute difference): \eqn{Matrix[i,j] = |attr[i] - attr[j]|}.

\code{"sqd.diff"}(squared difference): \eqn{Matrix[i,j] = (attr[i] - attr[j]=1)^2}.

\code{"sum"}(sum): \eqn{Matrix[i,j] = attr[i] + attr[j] = 1}.

\code{"product"}(product): \eqn{Matrix[i,j] = attr[i] * attr[j]}.

\code{"identity"}(identity coefficient): \eqn{Matrix[i,j] = 2 * (attr[i] * attr[j])/(attr[i]^2 + attr[j]^2}.

\code{"row"}(duplicating row actors' attributes: sender effect): \eqn{Matrix[i,j] = attr[i]}.

\code{"column"}(duplicating column actors' attributes: receiver effect): \eqn{Matrix[i,j] = attr[j]}.

\code{"max"}(maximam value): \eqn{Matrix[i,j] = max(attr[i], attr[j])}.

\code{"min"}(minimum value): \eqn{Matrix[i,j] = min(attr[i], attr[j])}.

\code{"mean"}(mean value): \eqn{Matrix[i,j] = mean(attr[i], attr[j])}.
}
\examples{
vec <- c(1,1,2,2,3,3)
names(vec) <- letters[1:6]
attr2mat(vec)
attr2mat(vec, "match.one")
attr2mat(vec, "match.one", diag.val = NULL)

}
