% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokerage.j.R
\name{brokerage.j}
\alias{brokerage.j}
\title{Positional analysis of j and k in the Brokerage relationship i-j-k.}
\usage{
brokerage.j(
  mat = mat,
  attr = attr,
  method = "abs.diff",
  alpha = 0.5,
  denom = "none"
)
}
\arguments{
\item{mat}{a matrix of a graph.}

\item{attr}{a vector of attributes.}

\item{method}{\code{"abs.diff"} or \code{"match"}. This is a intiger indicating the
method how to calculate the difference of the \code{attr} between i and j.
The default method \code{"abs.diff"} take the absolute
abs(attr[i]-attr[j]). \code{"match"} code 1 if attr[i]==attr[j] and 0
otherwise.}

\item{alpha}{a cut-off value to differentiate homophilous and heterophilious
ties. If the difference of \code{attr} values between i and j exceeds the
value of \code{alpha}, then the relationship between i and j is treated as
heterophilious relationship. Otherwise homophilous.}

\item{denom}{\code{"none"}(default), \code{"n"} or \code{"density"}. A
integer indicating how to standardize the counted number of times an actor
stands at the j position.}
}
\value{
a list of matrices in which each matrix shows the result of each
  typology.
}
\description{
\code{brokerage.j} and \code{brokerage.k} counts the number of times the
alter stands at the position in j and k, respectively, in the different
brokerage typology.
}
\details{
This function first differentiate the ties into homophilous and
heterophilious ties. It counts then how many times the alter stands in the j
position in the i-j-k relationship from the perspective of i. In the result
matrix the value of the cell M[i,j] indicates the number of times the alter j
stands in the j position. The count is standardized in accordance to the
method defined in \code{denom}. Brokerage analysis was first introduced in
Gould and Fernandez (1989). This function modifies the original concept in a
way that we construct the brokerage typology not based on the subgroup that
consists of actors engaging in the same issues, but the homophily and
heterophily ties.

In the brokerage relationship where i and j as well as j and k are connected
AND i and k is not connected,

Coordinator (wi) is the relationship where i->j and j->k is homophilous ties;

Itinerant broker (wo) is the one where i->j and j->k is heterophilious ties;

Representative (bio) is the one where i->j is a homophilous and j->k is a
heterophilious tie;

Gatekeeper (boi) is the one where i->j is a heterophilious and j->k is a
homophilous tie.

The fifth type, liaison, in the original literature (Gould and Fernandez
1989) does not come up in this function due to the aforementioned
modification: the heterophiles tie between i->j and j->k already implies the
same type to itinerant broker(bo).

The returned matrix M[i,j] is standardized if it is selected in
\code{"denom"}. \code{"n"} returns the value M[i,j]/n where n is number of
actors in the \code{mat}. \code{"density"} returns M[i,j]*density(mat).
}
\examples{
library(sna)
mat <- sna::rgraph(5, tprob = 0.5, mode = "graph")
dimnames(mat) <- list(LETTERS[1:5],LETTERS[1:5])
attr <- c(1,1,1,0,0)
gplot(mat, displaylabels = TRUE, vertex.col = attr, usearrows = FALSE)

brokerage.j(mat, attr, "match")
brokerage.j(mat, attr, "match", denom = "n")

}
\references{
Gould, Roger V. and Roberto M. Fernandez. (1989). "Structures of
  Mediation: A Formal Approach to Brokerage in Transaction Networks,"
  Sociological Methodology 19: 89-126.
}
