% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devilsf.R
\name{devilsf}
\alias{devilsf}
\title{Create a devil shift and angel shift matrix.}
\usage{
devilsf(
  mat = mat,
  pol = pol,
  score.std = "none",
  exp.method = "both",
  eval.method = "realized",
  alpha = NULL,
  score.norm = FALSE
)
}
\arguments{
\item{mat}{the matrix showing the influence checked by the row actors to
column actors.}

\item{pol}{the scale indicating the policy preference.}

\item{score.std}{string indicating the method of standardizing the policy
prefrence. score.std is "none" by default.}

\item{exp.method}{string indicating the method for calculating the degree of
overestimation.}

\item{eval.method}{string indicating the method to evaluate the
overestimation.}

\item{alpha}{cut-off value when binalize the devil shift and angel shift
matrix. alpha is NULL by default.}

\item{score.norm}{logical; should the engel- and devil shift score normalized
by the total number of influence checked in the matrix. score.norm is FALSE
by default.}
}
\description{
\code{devilsf} returns the devil shift and angel shift matrix.
}
\details{
Two devil shift and angel shift matrix is created from the two scale:(1) the
degree that the ego perceives the others more powerful than they actually are
(overestimation of powerfulness) and (2) the degree of opponents (devilness)
as well as proponents (angelness) from the perspective of respondents (egos).
Several options are available that need to be adjusted based on the
assumption of the researcher.
}
\examples{
mat <- matrix(c(0,0,0,1,1,1,
                1,1,1,0,0,0,
                1,0,0,1,0,0,
                1,1,1,0,0,0,
                0,0,0,1,1,1,
                1,0,0,1,0,0),6,6,byrow = TRUE)
namelabel <- c(paste(rep("A",3), 1:3, sep = ""),
               paste(rep("B",3), 1:3, sep = ""))
dimnames(mat) <- list(namelabel, namelabel)

pol <- c(1,1,0.6,0.4,0,0)
names(pol) <- namelabel
res <- devilsf(mat, pol,
              score.std = "none",
              exp.method ="both",
              eval.method = "realized",
              alpha = NULL,
              score.norm = F)

res$original.mat
res$exp.mat
res$eval.exp.mat
res$pol.dis
res$pol.agg
res$devil.mat
res$angel.mat

}
