% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscale.R
\name{gscale}
\alias{gscale}
\title{Scaling and centering of matrix.}
\usage{
gscale(mat, center = TRUE, scale = TRUE, diag = FALSE)
}
\arguments{
\item{mat}{a matrix to be transformed.}

\item{center}{a logical value indicating whether centering should be done.
The default value is \code{TRUE}.}

\item{scale}{a logical value indicating whether scaling should be done. The
default value is \code{TRUE}.}

\item{diag}{Whether the diagonal value is considered or not. The defalt value
is \code{FALSE}(not considered). With this defalut setting, the diagonal
values are set to be \code{NA}.}
}
\value{
a scaled matrix defined by the method.
}
\description{
\code{gscale} standardize the values in a matrix. This is a matrix version of
\code{scale} in the R package \code{base}.
}
\details{
This function first vectrizes the values in a matrix and then passes to the
\code{scale} function in R package  \code{base}. The returned values are then
converted into the original matrix format. If \code{center} is \code{TRUE},
then centering is done by subtracting the means (omitting \code{NA}s) of the
matrix from the original values in the matrix. If \code{scale} is
\code{TRUE}, then scaling is done by dividing the (centered) values in the
matrix by their standard deviations if \code{center} is \code{TRUE}, and the
root mean square otherwise. If \code{scale} is \code{FALSE}, no scaling is done.
If \code{diag} is \code{TURE}, the diagonal values are included for the
transformation.
}
\examples{
mat <- matrix(c(1,2,3,
               4,5,6,
               7,8,9),3,3,byrow = TRUE)
dimnames(mat) <- list(1:3, c("A","B","C"))

gscale(mat)
gscale(mat, diag = TRUE)
gscale(mat, center = TRUE, scale = FALSE, diag = TRUE)
gscale(mat, center = FALSE, scale = TRUE, diag = TRUE)

}
