% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo_hetr_mat.R
\name{homo_hetr_mat}
\alias{homo_hetr_mat}
\title{Differentiate homophilous and heterophilious graph.}
\usage{
homo_hetr_mat(mat, attr, alpha = 0.5)
}
\arguments{
\item{mat}{a matrix of a graph.}

\item{attr}{a vector of attributes.}

\item{alpha}{a cut-off value to differentiate homophilous and heterophilious
ties. If the difference of \code{attr} values between i and j exceeds the
value of \code{alpha}, then the relationship between i and j is treated as
heterophilious relationship. Otherwise homophilous.}
}
\value{
a list of matrices: a homopilious matrix and a heterophilious matrix.
}
\description{
\code{homo_hetr_mat} takes a matrix and returns a list of matrices in which
homophilous and heterophilious ties are differentiated.
}
\examples{
library(sna)
mat <- sna::rgraph(5)
dimnames(mat) <- list(LETTERS[1:5], LETTERS[1:5])
attr <- c(0,0,0,1,1)

homo_hetr_mat(mat, attr)

}
