% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvif.R
\name{netvif}
\alias{netvif}
\title{Calculate vif for network regression.}
\usage{
netvif(y, x, mode = "digraph", diag = FALSE, fun = "netlm")
}
\arguments{
\item{y}{dependent network variable.}

\item{x}{stack of independent network variables (array or list object).}

\item{mode}{string indicating the type of graph being evaluated.
\code{"digraph"}(default: directed graph) or \code{"graph"}(undirected
graph) .}

\item{diag}{logical. Whether the diagonal value is considered or not
(default).}

\item{fun}{intiger indicating the regression model: \code{"netlm"}(default)
or \code{"netlogit"}.}
}
\value{
a vector indicating vifs.
}
\description{
\code{netvif} vectorizes the object for the QAP regression and
passes to the function \code{vif} in the package \code{car} to calculate the vif.
}
\details{
This is a utility function for selecting the independent variables when
conducting QAP regression.
}
\examples{
# Create some input graphs
x <- sna::rgraph(20,4)

# Create a response structure
y <- x[1,,]+4*x[2,,]+2*x[3,,]   #Note that the fourth graph is unrelated

# calculate the vifs when using these items
netvif(y, x, fun = "netlm")
}
\seealso{
\code{\link[car]{vif}}, \code{\link[sna]{netlm}},
  \code{\link[sna]{netlogit}}.
}
