% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmperm.grpd.R
\name{rmperm.grpd}
\alias{rmperm.grpd}
\title{Randomly Permute the Rows and Columns of an Input Matrix by group of actors.}
\usage{
rmperm.grpd(mat, grp.ns = NULL)
}
\arguments{
\item{mat}{a matrix, or stack thereof. For a stack of matrices, it should be either list format or array in which the first dimension indicates the layer.}

\item{grp.ns}{a vector indicates the number of actors in each group. The default is \code{NULL}.}
}
\value{
a permuted matrix (or  matrices).
}
\description{
Given an input matrix (or stack thereof), \code{rmperm.grpd} performs a
(random) simultaneous row/column permutation of the input data BY EACH GROUP.
That means permutation does not occur among the different groups.
}
\details{
This is the extention of the \code{\link[sna]{rmperm}} for the matrix
permtation. If no group was defined in \code{grp.ns}, it performs the normal
matrix permutation. If groups are defined, it perfom the matrix permutation
by each group seperately and then convert it into a single (or stacked)
matrix. This function helps for QAP test in which the off-diagonal matrix is
set to be \code{NA}, for example.
}
\examples{
library(sna)
# single group/single net(normal permutation)
m <- sna::rgraph(4)
rmperm.grpd(m)

# single group/multiplex(normal permutation)
m <- rgraph(4,2)
m[1,,];m[2,,]
res <- rmperm.grpd(m)
res[1,,];res[2,,]

# multi group/single net
m1 <- sna::rgraph(4)
m2 <- sna::rgraph(3)
m3 <- sna::rgraph(3)
spr.mat <- list2mat(list(m1,m2,m3))
spr.mat
rmperm.grpd(spr.mat, grp.ns = c(4,3,3))

# multi group/multiplex
spr.mat2 <- spr.mat
spr.mat2[3,2]<-1;spr.mat2[1,3]<-1
sprmats <- list(spr.mat,spr.mat2)
res <- rmperm.grpd(sprmats, grp.ns = c(nrow(m1),nrow(m2),nrow(m3)))
res[1,,];res[2,,]

}
